// admisi/src/types/CalonMahasiswaTypes.ts
export type CalonMahasiswaType = {
  NomorDaftar: string;
  NamaCamhs: string;
  ProdiId: 41;
  NamaProdi: string;
  WaktuKuliah: string;
  JenisUjian: string;
  StatusDokumen: string;
  JenisDaftar: string;
  StatusKeuangan: string;
  KeteranganPembayaran: string;
  KeteranganDokumen: string;
  PemilikRekeningKeuangan: string;
  BankBuktiBayar: string;
  LinkBuktiBayar: string;
  JenjangCamhs: string;
  PasPhotoLink: string;
  IsBeasiswa: string;
  VaCamhs: string;
  JalurDaftar: string;
  NoHpCamhs: string;
  JkCamhs: string;
  ProgressCamhs: string;
  EmailCamhs: string;
  FormulirCicilan: string;
  BesiswaId: number;
  BeasiswaNama: string;
  StatusPenerimaan: string;
  PeriodeCamhs: string;
  GelombangCamhs: string;
  DetailgelombangId: number;
  MkAgama: string;
  PeringkatId: number;
  PeringkatNama: string;
  SizeAlmamater: string;
  AsalSekolah: "SMAS MONDIAL";
  NpmCamhs: string;
  AgamaCamhs: string;
  AverageNilaiUsm: number;
  IsWawancara: string;
  StatusKeuanganDaftarUlang: string;
  TanggalKeuanganDaftarUlang: string;
  ApprovalTagihanData: ApprovalTagihanDataType;
};

export type CalonMahasiswaRPLType = {
  NomorDaftar: string;
  NamaCamhs: string;
  JkCamhs: string;
  ProdiId: number;
  NamaProdi: string;
  JenjangCamhs: string;
  NoTelp: string;
  NoWa: string;
  EmailCamhs: string;
  LineCamhs: string;
  TanggalDaftar: string;
  BiodataCamhs: BiodataCalonMahasiswaRPLType;
  DataOrtuCamhs: DataOrtuCalonMahasiswaRPLType;
  DokumenData: DokumenDataCalonMahasiswaRPLType;
  LamaPengalaman: string;
  LinkBuktiPengalaman: string;
  StatusPengajuanRpl: string;
  TanggalSubmitRpl: string;
  StatusCamhs: string;
  StatusBiodataCamhs: string;
  StatusDokumenCamhs: string;
  StatusKeuanganCamhs: string;
};

export type BiodataCalonMahasiswaRPLType = {
  NomorDaftar: string;
  NikCamhs: string;
  KepalaKeluargaCamhs: string;
  NoKkCamhs: string;
  JurusanCamhs: string;
  AsalPendidikanTerakhir: string;
  JalurPendaftaran: string;
  NomorIjazah: string;
  WaktuKuliah: string;
  TanggalLahir: string;
  TempatLahir: string;
  NisnCamhs: string;
  NpwpCamhs: string;
  AgamaId: number;
  AgamaCamhs: string;
  KewarganegaraanId: number;
  KewarganegaraanCamhs: string;
  AlamatCamhs: string;
  NamaDusunCamhs: string;
  NoRtCamhs: string;
  NoRwCamhs: string;
  KelurahanCamhs: string;
  KecamatanCamhs: string;
  KodePosCamhs: string;
  NamaInstansi: string;
  AlamatInstansi: string;
  StatusInstansi: string;
  TahunMulaiKerja: string;
  Jabatan: string;
  LinkCv: string;
  SumberBiaya: string;
  BidangKeahlian: string;
  BiayaPendaftaran: string;
};

export type DataOrtuCalonMahasiswaRPLType = {
  NamaAyah: string;
  NikAyah: string;
  PekerjaanAyahId: string;
  NamaPekerjaanAyah: string;
  PendidikanAyah: string;
  TanggalLahirAyah: string;
  IsMeninggalAyah: string;
  NoHpAyah: string;
  NamaIbu: string;
  NikIbu: string;
  PekerjaanIbuId: string;
  NamaPekerjaanIbu: string;
  PendidikanIbu: string;
  TanggalLahirIbu: string;
  IsMeninggalIbu: string;
  NoHpIbu: string;
  NamaWali: string;
  NikWali: string;
  PekerjaanWaliId: string;
  NamaPekerjaanWali: string;
  PendidikanWali: string;
  TanggalLahirWali: string;
  IsMeninggalWali: string;
  NoHpWali: string;
  AlamatOrtu: string;
  AlamatWali: string;
};

export type DokumenDataCalonMahasiswaRPLType = {
  AktaLahirNama: string;
  AktaLahirLink: string;
  KkNama: string;
  KkLink: string;
  PasPhotoNama: string;
  PasPhotoLink: string;
  KtpNama: string;
  KtpLink: string;
  IjazahNama: string;
  IjazahLink: string;
  KartuPelajarNama: string;
  KartuPelajarLink: string;
  NpwpNama: string;
  NpwpLink: string;
  BankPembayaran: string;
  PemilikRekening: string;
  BuktiBayarNama: string;
  BuktiBayarLink: string;
};

export type ApprovalTagihanDataType = {
  StatusValidasi: string;
  QrCode: string;
  NipApproval: string;
  UserApproval: string;
  TglApproval: string;
};